Tortoise_ORM/requirements.txt
aerich==0.7.2
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
dictdiffer==0.9.0
fastapi==0.115.5
idna==3.10
iso8601==2.1.0
pydantic==2.9.2
pydantic_core==2.23.4
pypika-tortoise==0.2.1
pytz==2024.2
sniffio==1.3.1
starlette==0.41.2
tomlkit==0.13.2
tortoise-orm==0.21.7
typing_extensions==4.12.2
